site_parameters.xml read by StudySiteDataContainer
Noticed during the course of implementing this that FlowDirectionNetwork
cannot be initialised for grids the size I need, as the indices go like
N^2 where N is the number of /cells/ in the grid. In a 500*500 grid,
this means the indices need to to to 500^4 = 62,500,000,000 which is
larger than the largest int Java can hold. Will need to re-implement
soil moisture calculator