Change LcsUpdateMsg to return complete environmental state
The new LcsUpdateMsg will include information about the complete state
of the grid cell, not just its current and target states and time in state.
This will enable it to communicate if a land-cover state transition has had
an impact on the cells environmental conditions as a side effect of the
transition.

The old implementation is now called `LegacyLcsUpdateMsg`